update Xcode projects to compile on master
The old header search path entry for the Boost submodule needed to be removed and the OS X projects needed to link to IOKit. Looks like Xcode now strictly enforces some sort of alphabetical ordering in the plist so there are a bunch of changes it made that just reordered properties :0\

It is worth noting that the RTT example for iOS may need some further attention as it didn't display anything.